\name{makesigmasemiPD}
\alias{makesigmasemiPD}
\title{positive semidefinite}
\description{
Compute the positive semidefinite
}
\usage{
makesigmasemiPD(geno)
}
\arguments{
  \item{geno}{a m x n numeric matrix}
}
\details{
Compute the positive semidefinite
}
\references{
  Leveraging allelic heterogeneity to increase power of association testing
(https://www.biorxiv.org/content/10.1101/498360v1)
}
\examples{
data(testdata)
makeSigmasemiPD(geno)
}
\author{Junghyun Jung}


